added blocklist.py and the /logout endpoint

+++ resources/item.py
@@ -1,7 +1,8 @@
+from email import message
 from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -14,6 +15,7 @@ blp = Blueprint("Items", __name__, description="Operations on items.")
 @blp.route('/item/<int:item_id>')
 class Item(MethodView):
 
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
@@ -21,6 +23,11 @@ class Item(MethodView):
 
     @jwt_required()
     def delete(self, item_id):
+        """Delete the item with the `item_id`, only if the user has admin identity."""
+        jwt = get_jwt()
+        if not jwt.get('is_admin'):
+            abort(401, message='Admin privilege required.')
+        
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -49,6 +56,7 @@ class Item(MethodView):
 
 @blp.route('/item')
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
